transport: correct format string when printing logging message
we print the stream id in the logging messages, but in this case,
we forgot to pass `stream` to `log::debug()`. but the placeholder
for `stream` was added. if the underlying fmtlib actually formats
the argument with the format string, it would throw.

fortunately, we don't enable debug level logging often, guess that's
why we haven't spotted this issue yet.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #14620