sstables: Automatically close exhausted SSTable readers in cleanup
Add a reader that will automatically close the underlying sstable
reader if fast forward is called with a range past the range
spanned by the SSTable. This is only to be used in the context
of fast forward calls in cleanup, as combined reader in full
scans can proactively close the readers that returned EOS.

Regular reads that go through cache enable fast forwarding to
position range, therefore won't enable auto-closed reader.
Compactions don't enable any kind of forward, and they won't
have it enabled either.

The overhead is minimal, with cleanup being able to reach the
same 38MB/s as before this patch.

Refs #12998.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>