cql3: grammar: don't default-construct expressions
Use uninitialized<expression> for that. Since it's heavily used,
alias it as "uexpression".

To prevent uninitialized<> from leaking into the rest of the
system, change do_with_parser() to unwrap it. We add an
unwrap_uninitialized_t template type alias for that.

Lots of std::move()s are sprinkled around to make things compile,
as uninitialized<T> refuses to convert to T without them.