cql3: grammar: improve uninitialized<> flexibility
uninitialized<> is used to work around the parser generator's propensity
to default-construct return values by supplying a default constructor
to otherwise non-default-constructible types. Make it easier to initialize
it not only from the wrapped type, but also from types convertible to
the wrapped type.

This is useful to initialize an uninitialized<expression> from an
expression element (say a binary_operator), without an explicit
conversion.